Add files via upload
this is a work in progress plugin. note that im still working on it. the region is North america. 

i cannot promise to get desired cheats. this is beta 1.0 wich should work. as previewed on youtube. 

this is just a hobby project im sometimes doing. as the same as the shorts on my channel. note that **EVERYTHING i do** is considered as my art . i consider myself as an artist. so i cannot say for sure that the stuff im doing gets updated quickly or even generally updated.